azure_test/requirements.txt
azure-functions
azure-storage-blob
azure-search-documents
azure-identity
pydantic
InquirerPy
langchain